Update to retry entire package upload
Upload was seeming to fail even after the page surely had returned. I
was seeing a NoSuchElementException thrown, and handled, but the
element wasn't there so after it retried, the implicit timeout kicked
in and a NoSuchElementException was thrown but not handled.

This approach should ensure a full retry including reloading the
upload page.